+++ tablet_qt/lib/uifunc.cpp
@@ -20,12 +20,10 @@
 
 // #define DEBUG_ICON_LOAD
 // #define DEBUG_SCROLL_GESTURES
-// #define DEBUG_MIN_SIZE_FOR_TITLE
 
 #include "uifunc.h"
 
 #include <QAbstractItemView>
-#include <QApplication>
 #include <QBrush>
 #include <QDebug>
 #include <QDesktopServices>
@@ -50,7 +48,6 @@
 
 #include "common/colourdefs.h"
 #include "common/languages.h"
-#include "common/platform.h"
 #include "common/textconst.h"
 #include "common/uiconst.h"
 #include "core/camcopsapp.h"
@@ -656,73 +653,6 @@ void makeItemViewScrollSmoothly(QObject* object)
 // Sizing
 // ============================================================================
 
-QSize minimumSizeForTitle(const QDialog* dialog, const bool include_app_name)
-{
-    if (!dialog) {
-        return QSize();
-    }
-    // +---------------------------------------------+
-    // | ICON  TITLETEXT - APPTITLE    WINDOWBUTTONS |
-    // |                                             |
-    // | contents                                    |
-    // +---------------------------------------------+
-
-    // https://doc.qt.io/qt-6.5/qwidget.html#windowTitle-prop
-    const QString window_title = dialog->windowTitle();
-    const QString app_name = QApplication::applicationDisplayName();
-    QString full_title = window_title;
-    if (include_app_name && !platform::PLATFORM_TABLET) {
-        // Qt for Android doesn't append this suffix.
-        // It does for Linux and Windows.
-        const QString title_suffix = QString(" — %1").arg(app_name);
-        full_title += title_suffix;
-    }
-    const QFont title_font = QApplication::font("QWorkspaceTitleBar");
-    const QFontMetrics fm(title_font);
-    const int title_w = fm.boundingRect(full_title).width();
-    // ... "_w" means width
-
-    // dialog->ensurePolished();
-    // const QSize frame_size = dialog->frameSize();
-    // const QSize content_size = dialog->size();
-    // ... problem was that both are QSize(640, 480) upon creation
-    // ... even if ensurePolished() is called first
-    // const QSize frame_extra = frame_size - content_size;
-
-    // How to count the number of icons shown on a window? ***
-    // - Android: 0
-    // - Linux: presumably may vary with window manager, but 4 is typical under
-    //   XFCE (1 icon on left, 3 [rollup/maximize/close] on right), but need a
-    //   bit more for spacing; 6 works better (at 24 pixels width per icon)
-    // - Windows: also 4 (icon left, minimize/maximize/close on right)
-    const int n_icons = platform::PLATFORM_TABLET ? 0 : 6;
-
-    // How to read the size (esp. width) of a window icon? ***
-    // const int icon_w = frame_extra.height();
-    // ... on the basis that window icons are square!
-    // ... but the problem is that frame size may as yet be zero
-    const int icon_w = 24;
-
-    const int final_w = title_w + n_icons * icon_w;
-    const QSize dialog_min_size = dialog->minimumSize();
-    QSize size(dialog_min_size);
-    size.setWidth(qMax(size.width(), final_w));
-    size.setWidth(qMin(size.width(), dialog->maximumWidth()));
-#ifdef DEBUG_MIN_SIZE_FOR_TITLE
-    qDebug().nospace() << Q_FUNC_INFO << "window_title = " << window_title
-                       << ", app_name = " << app_name
-                       << ", full_title = " << full_title
-                       << ", title_font = " << title_font << ", title_w = "
-                       << title_w
-                       // << ", frame_size = " << frame_size
-                       // << ", content_size = " << content_size
-                       // << ", frame_extra = " << frame_extra
-                       << ", n_icons = " << n_icons << ", icon_w = " << icon_w
-                       << ", dialog_min_size = " << dialog_min_size
-                       << ", size = " << size;
-#endif
-    return size;
-}
 
 QScreen* screen()
 {